fixed delete table test
